scale displayed image to save memory
When using this helper for a lot of pictures that are not mobile
optimized won get an out of memory very fast.
The trick is to display a scaled down image.

By default, images have a MAX_DISPLAY_SIZE of 100
so if you load an image that is 2000x2000 pixel big, it will be shown
as a 100x100 picture.

The setUrlDrawable methods also offer the ability to set an other value
if desired