Use a dead, live, and disk cache. In that order.
Live cache has soft references to ZombieDrawables.
When a Zombie is finalized, it goes into the dead cache.
When requesting a url, the dead cache is checked. If found,
the drawable gets put back into the live cache.

This fixes the GC spam issue on list scrolling, when
drawables quickly leave and reenter the object graph.

wip

remove hacks